GM
GM now has a cancel button.

+++ src/museumtimetracking/gui/views/root/guildManager/guildManagerOverview/GuildManagerOverviewController.java
@@ -76,6 +76,8 @@ public class GuildManagerOverviewController implements Initializable {
     private static final String EDIT_BUTTON_TEXT = "Rediger";
     private static final String SAVE_BUTTON_TEXT = "Gem";
     private static final String NEW_GUILD_MANAGER_TEXT = "Ny Tovholder";
+    private static final String CANCEL_BUTTON_TEXT = "Annuller";
+    private static final String DELETE_BUTTON_TEXT = "Slet";
 
     private final ModalFactory modalFactory;
 
@@ -163,21 +165,25 @@ public class GuildManagerOverviewController implements Initializable {
     @FXML
     private void handleDeleteButton() {
         GM managerToDelete = lstManagers.getSelectionModel().getSelectedItem();
-        Alert deleteAlert = AlertFactory.createDeleteAlert();
-        deleteAlert.showAndWait().ifPresent(type -> {
-            //If user clicks first button
-            if (type == deleteAlert.getButtonTypes().get(0)) {
-                try {
-                    guildManagerModel.deleteGuildManager(managerToDelete);
-                } catch (DALException ex) {
-                    ExceptionDisplayer.display(ex);
+        if (btnDelete.getText().equals(DELETE_BUTTON_TEXT)) {
+            Alert deleteAlert = AlertFactory.createDeleteAlert();
+            deleteAlert.showAndWait().ifPresent(type -> {
+                //If user clicks first button
+                if (type == deleteAlert.getButtonTypes().get(0)) {
+                    try {
+                        guildManagerModel.deleteGuildManager(managerToDelete);
+                    } catch (DALException ex) {
+                        ExceptionDisplayer.display(ex);
+                    }
                 }
-            }
-        });
-
+            });
+        } else if (btnDelete.getText().equals(CANCEL_BUTTON_TEXT)) {
+            setShowEditability(false);
+        }
         setButtonTextToViewMode();
         setSetsToNull();
         lstManagers.refresh();
+        updateInformation(managerToDelete);
     }
 
     /**
@@ -310,8 +316,9 @@ public class GuildManagerOverviewController implements Initializable {
      */
     private void setButtonTextToEditMode() {
         btnNewGuildManager.setText(ADD_GUILD_BUTTON_TEXT);
-        btnDelete.setDisable(true);
-        btnDelete.setVisible(false);
+        btnDelete.setText(CANCEL_BUTTON_TEXT);
+//        btnDelete.setDisable(true);
+//        btnDelete.setVisible(false);
         btnEdit.setText(SAVE_BUTTON_TEXT);
         lstManagers.setDisable(true);
         btnArchiveManager.setDisable(true);
@@ -325,8 +332,9 @@ public class GuildManagerOverviewController implements Initializable {
      */
     private void setButtonTextToViewMode() {
         btnNewGuildManager.setText(NEW_GUILD_MANAGER_TEXT);
-        btnDelete.setDisable(false);
-        btnDelete.setVisible(true);
+        btnDelete.setText(DELETE_BUTTON_TEXT);
+//        btnDelete.setDisable(false);
+//        btnDelete.setVisible(true);
         btnEdit.setText(EDIT_BUTTON_TEXT);
         lstManagers.setDisable(false);
         btnArchiveManager.setDisable(false);
